+++ package.json
@@ -2,7 +2,8 @@
   "name": "storybook-styled-components",
   "version": "1.0.1",
   "description": "",
-  "main": "dist/decorator.js",
+  "main": "dist/index.js",
+  "jsnext:main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "babel src --out-dir dist/",
